pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quantai"
version = "0.1.0"
description = "Quant AI trading system — OpenBB · QuantAgent · MiroFish · QuantMuse"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Data layer
    "openbb>=4.3.0",
    "openbb-yfinance>=0.2.0",

    # Quant / ML
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",

    # LLM / Agents
    "anthropic>=0.28.0",
    "langchain>=0.2.0",
    "langgraph>=0.1.0",

    # HTTP / messaging
    "httpx>=0.27.0",
    "pyzmq>=26.0.0",
    "pydantic>=2.7.0",

    # Dashboard
    "streamlit>=1.35.0",
    "plotly>=5.22.0",

    # Utils
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["data", "signals", "simulation", "execution", "dashboard"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
